refactor(control_algorithms): Split astar files into different algorithms
